Extract shared physics into physics.py, add load_model_from_checkpoint
- runner.py: Add load_model_from_checkpoint() for cross-stage handoff
  (partial orbax restore of tensors only)
- z2_vison/physics.py: Shared Hamiltonian, model, plaquette observables,
  vison insertion. ground_state.py and dynamics.py import from it.
- z2_vison_higgs/physics.py: Shared Higgs Hamiltonian, model, plaquette
  observables, interior vison pair. Scripts import from it.
- Remove save_model_state/load_model_state from scripts (runner handles)
